pyproject.toml
[project]
name = "apple-doc-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "faiss-cpu>=1.10.0",
    "langchain>=0.3.21",
    "langchain-community>=0.3.20",
    "langchain-huggingface>=0.1.2",
    "llama-cpp-python>=0.3.8",
    "nltk>=3.9.1",
    "pip-system-certs>=4.0",
    "requests>=2.32.3",
    "selenium>=4.30.0",
    "sentence-transformers>=4.0.1",
    "tqdm>=4.67.1",
    "unstructured>=0.17.2",
    "webdriver-manager>=4.0.2",
]
